Force integer comparison for account flags
The account_readonly and account_admin columns of the login_account
table are defined as TINYINT in the MariaDB database. Some time
between 2025-05-13 and 2025-06-03 GoDaddy made an unannounced change
to either the PHP configuration or the MariaDB installation which
caused queries which had been returning values from those columns
as integers to suddenly return string values instead, which broke
the code which determines which permissions the account has. The
code was changed to coerce the strings back to integers before
comparing them to constant values.